+++ tests/mod.rs
@@ -45,6 +45,12 @@ fn with_inner_test_attribute_and_test_args(x: i8, y: i8) {
   assert_eq!(y, -4);
 }
 
+#[test_log::test(test_case::test_case(-2, -4; "my test name"))]
+fn with_inner_test_attribute_and_test_args_and_name(x: i8, y: i8) {
+  assert_eq!(x, -2);
+  assert_eq!(y, -4);
+}
+
 #[should_panic]
 #[test_log::test(test_case::test_case(-2, -4))]
 fn with_inner_test_attribute_and_test_args_and_panic(x: i8, _y: i8) {